No-op refactors to vue source code
