+++ canvases/dark-editor-vue/src/App.vue
@@ -1,6 +1,7 @@
-<template>
-  <MainView />
-</template>
 <script setup lang="ts">
 import MainView from './views/MainView.vue'
 </script>
+
+<template>
+  <MainView />
+</template>
