+++ canvases/dark-editor-vue/src/components/common/AutoSizeTextarea.vue
@@ -1,10 +1,7 @@
-<template>
-  <textarea ref="content" class="text-white bg-transparent w-full outline-none" v-model="localResponse"></textarea>
-</template>
-
-<script setup lang="ts">
+<!-- <script setup lang="ts">
 import { defineProps, ref, onMounted, watch, defineEmits, computed } from 'vue'
 import autosize from 'autosize'
+
 const props = defineProps({
   response: {
     type: String,
@@ -13,6 +10,7 @@ const props = defineProps({
 })
 
 const content = ref(null)
+
 const localResponse = computed({
   get: () => props.response,
   // update the parent component whenever the textarea value changes
@@ -41,3 +39,11 @@ watch(
   }
 )
 </script>
+
+<template>
+  <textarea
+    ref="content"
+    class="text-white bg-transparent w-full outline-none"
+    v-model="localResponse"
+  ></textarea>
+</template> -->
